restructure components folder
