--- client/src/components/Features/Features.jsx
@@ -1,15 +1,0 @@
-import classes from './Features.module.css';
-
-function Features() {
-  return (
-    <main className={classes.features}>
-      <div>
-        <div>
-          <div></div>
-        </div>
-      </div>
-    </main>
-  );
-}
-
-export default Features;
